requirements.txt
streamlit==1.28.1
openai==1.3.7
pinecone-client==2.2.4
langchain==0.0.335
langchain-openai==0.0.2
langchain-pinecone==0.0.1
python-dotenv==1.0.0
pandas==2.1.3

# Google Drive Integration (Required)
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Removed local file system dependencies
# watchdog==3.0.0  # No longer needed for file watching
# markdown==3.5.1  # No longer needed for local markdown processing

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
